LayoutTests/media/media-source/media-source-real-webm-remove.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-webm-remove</title>
    <script src="webm-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        if (!MediaSource.isTypeSupported(WebM.VIDEO_TYPE)) {
            consoleWrite('SKIPPED: ' + WebM.VIDEO_TYPE + ' is not supported');
            return;
        }

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        sourceBuffer = source.addSourceBuffer(WebM.VIDEO_TYPE);

        sourceBuffer.appendBuffer(WebM.initSegment({ tracks: [{ id: 1, type: 'video' }] }));
        await waitFor(sourceBuffer, 'updateend');

        // 8 seconds, sync at 0 and 4
        sourceBuffer.appendBuffer(WebM.mediaSegment({
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples: [
                { duration: 1000, isSync: true },
                { duration: 1000, isSync: false },
                { duration: 1000, isSync: false },
                { duration: 1000, isSync: false },
                { duration: 1000, isSync: true },
                { duration: 1000, isSync: false },
                { duration: 1000, isSync: false },
                { duration: 1000, isSync: false },
            ]}]
        }));
        await waitFor(sourceBuffer, 'updateend');

        testExpected('sourceBuffer.buffered.length', 1);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.end(0)', 8, 0.05);

        // Remove [0, 2) — should extend to sync frame at 4
        sourceBuffer.remove(0, 2);
        await waitFor(sourceBuffer, 'updateend');

        testExpected('sourceBuffer.buffered.length', 1);
        testExpected('sourceBuffer.buffered.start(0)', 4);
    }

    window.addEventListener('load', event => {
        runTest().then(endTest).catch(failTest);
    });
    </script>
</head>
<body>
    <video></video>
</body>
</html>
