LayoutTests/media/media-source/media-source-real-webm-remove-expected.txt


EVENT(sourceopen)
EVENT(updateend)
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.end(0) == '8') OK
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '4') OK
END OF TEST
